# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TournamentStatus =
  (typeof TournamentStatus)[keyof typeof TournamentStatus];

export const TournamentStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface Participant {
  id: number;
  tournamentId: number;
  name: string;
  /** @nullable */
  seed: number | null;
  createdAt: string;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  pending: "pending",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface Match {
  id: number;
  tournamentId: number;
  round: number;
  matchNumber: number;
  /** @nullable */
  participant1Id: number | null;
  /** @nullable */
  participant1Name: string | null;
  /** @nullable */
  participant2Id: number | null;
  /** @nullable */
  participant2Name: string | null;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  winnerName: string | null;
  /** @nullable */
  score1: string | null;
  /** @nullable */
  score2: string | null;
  status: MatchStatus;
  createdAt: string;
  updatedAt: string;
}

export interface Tournament {
  id: number;
  name: string;
  /** 1v1, 2v2, 3v3, etc. */
  type: string;
  status: TournamentStatus;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  winnerName: string | null;
  createdAt: string;
  updatedAt: string;
  participants: Participant[];
  matches: Match[];
}

export interface CreateTournamentBody {
  name: string;
  type: string;
}

export type UpdateTournamentBodyStatus =
  (typeof UpdateTournamentBodyStatus)[keyof typeof UpdateTournamentBodyStatus];

export const UpdateTournamentBodyStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  archived: "archived",
} as const;

export interface UpdateTournamentBody {
  name?: string;
  type?: string;
  status?: UpdateTournamentBodyStatus;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
}

export interface AddParticipantBody {
  tournamentId: number;
  name: string;
  /** @nullable */
  seed?: number | null;
}

export type UpdateMatchBodyStatus =
  (typeof UpdateMatchBodyStatus)[keyof typeof UpdateMatchBodyStatus];

export const UpdateMatchBodyStatus = {
  pending: "pending",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface UpdateMatchBody {
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
  /** @nullable */
  score1?: string | null;
  /** @nullable */
  score2?: string | null;
  status?: UpdateMatchBodyStatus;
}

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNewsBody {
  title: string;
  content: string;
}

export interface UpdateNewsBody {
  title?: string;
  content?: string;
}