# lib/api-zod/src/generated/types/addParticipantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AddParticipantBody {
  tournamentId: number;
  name: string;
  /** @nullable */
  seed?: number | null;
}