# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from "./matchStatus";

export interface Match {
  id: number;
  tournamentId: number;
  round: number;
  matchNumber: number;
  /** @nullable */
  participant1Id: number | null;
  /** @nullable */
  participant1Name: string | null;
  /** @nullable */
  participant2Id: number | null;
  /** @nullable */
  participant2Name: string | null;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  winnerName: string | null;
  /** @nullable */
  score1: string | null;
  /** @nullable */
  score2: string | null;
  status: MatchStatus;
  createdAt: Date;
  updatedAt: Date;
}