# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}