# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Participant {
  id: number;
  tournamentId: number;
  name: string;
  /** @nullable */
  seed: number | null;
  createdAt: Date;
}