# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Match } from "./match";
import type { Participant } from "./participant";
import type { TournamentStatus } from "./tournamentStatus";

export interface Tournament {
  id: number;
  name: string;
  /** 1v1, 2v2, 3v3, etc. */
  type: string;
  status: TournamentStatus;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  winnerName: string | null;
  createdAt: Date;
  updatedAt: Date;
  participants: Participant[];
  matches: Match[];
}