# lib/api-zod/src/generated/types/updateMatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMatchBodyStatus } from "./updateMatchBodyStatus";

export interface UpdateMatchBody {
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
  /** @nullable */
  score1?: string | null;
  /** @nullable */
  score2?: string | null;
  status?: UpdateMatchBodyStatus;
}