# lib/api-zod/src/generated/types/updateTournamentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for خلية البطولات - Tournament Platform
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTournamentBodyStatus } from "./updateTournamentBodyStatus";

export interface UpdateTournamentBody {
  name?: string;
  type?: string;
  status?: UpdateTournamentBodyStatus;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
}